requirements.txt
customtkinter
pyotp
requests
websocket-client
python-dotenv
